contests/joi2007ho/c/main_test.go
package main

import (
	"fmt"
	"os"
	"path/filepath"
	"testing"
)

func TestSolution(t *testing.T) {
	t.Parallel()
	type in struct {
		n int
		a [][2]int
	}
	tests := []struct {
		in   in
		want int
	}{
		{in: in{n: 10, a: [][2]int{{9, 4}, {4, 3}, {1, 1}, {4, 2}, {2, 4}, {5, 8}, {4, 0}, {5, 3}, {0, 5}, {5, 2}}}, want: 10},
	}
	for i, tt := range tests {
		i, tt := i, tt
		t.Run(fmt.Sprint(i), func(t *testing.T) {
			t.Parallel()
			got := solution(tt.in.n, tt.in.a)
			if got != tt.want {
				t.Fatalf("in: %v got: %v want: %v", tt.in, got, tt.want)
			}
		})
	}
}

func Example_main() {
	fd, _ := os.Open(filepath.Join("testdata", "input.txt"))
	orgStdin := os.Stdin
	os.Stdin = fd
	defer func() {
		os.Stdin = orgStdin
		fd.Close()
	}()

	main()

	// Output: 10
}
